Retrieving airdate along with stage data
